+++ src/overlay/FlowControlCapacity.h
@@ -83,7 +83,7 @@ class FlowControlByteCapacity : public FlowControlCapacity
     virtual void releaseOutboundCapacity(StellarMessage const& msg) override;
     bool canRead() const override;
 #ifdef ENABLE_NEXT_PROTOCOL_VERSION_UNSAFE_FOR_PRODUCTION
-    void handleTxSizeIncrease(uint64_t increase);
+    void handleTxSizeIncrease(uint32_t increase);
 #endif
 };
